Move routine tests out of da.tests.test_array_core
